.half() conversion

+++ GraphRec-WWW19/run_GraphRec_example.py
@@ -201,13 +201,13 @@ def main():
     num_embeddings, embedding_dim = max_item_id + 1, embed_dim
     embed_weight = torch.zeros(num_embeddings, embedding_dim, dtype=torch.float16).to(device)
     torch.nn.init.normal_(embed_weight)
-    uv2e = nn.Embedding(num_embeddings, embedding_dim, _weight=embed_weight).to(device)
+    uv2e = nn.Embedding(num_embeddings, embedding_dim, _weight=embed_weight).to(device, dtype=torch.float16)
     # log.warn("Now next large type conversion... it may crash")
     # uv2e32 = uv2e.float()
     log.info("uv2e {}x{} embeddings initialized".format(max_item_id + 1, embed_dim))
     log.warn("Use embeddings more effectively to not waist memory usage")
 
-    r2e = nn.Embedding(num_ratings, embed_dim).to(device)
+    r2e = nn.Embedding(num_ratings, embed_dim).to(device, dtype=torch.float16)
     log.info("r2e {}x{} embeddings initialized".format(num_ratings, embed_dim))
 
     # user feature
